+++ src/painter/pen_brush_gui.cpp
@@ -85,7 +85,6 @@ void pen_brush_gui::init_layout()
         QString module_name("pbgui");
         m_solid_line = new QRadioButton("Solid Line", this);
         m_solid_line ->setObjectName(module_name+"/"+"rbSolidLine");
-        
         //set_log_name(m_solid_line);
         
         //FIXME automate this with following function
